Checking if the referent has been cleared (#698)
Some runtimes clear reference to a different value other than null. This PR adds a method to the ReferenceGlue trait to check if a reference is cleared and replac the checks in mmtk-core to call this method instead. Therefore, bindings can overwrite the method `is_referent_cleared` to do the check properly.