Build MMTk on macOS (#719)
This PR allows mmtk-core to build on macOS (x86_64). This PR finishes a task in https://github.com/mmtk/mmtk-core/issues/206. This PR does not guarantee that MMTk works properly on macOS. It only makes sure that we can build and run our unit tests on macOS. This should be a good first step for any following work on macOS.

Changes:
* Use `malloc_size` instead of `malloc_usable_size` on macOS
* Use `MAP_FIXED` instead of `MAP_FIXED_NOREPLACE` on macOS
* Remove warnings when building `scheduler::affinity` on macOS
* Make a few tests only run on linux (related with thread affinity and mmap)
* Test our pre-code-review test on `macos-12` and `x86_64-apple-darwin`