Fix slow vec allocation during start-up. (#737)
Use alloc::alloc_zeroed to manually allocate the buffer for the vector of `SpaceDescriptor(usize)` to elide the storing of zeroes to its elements during start-up.

This is one of the two parts for solving the problem of slow start-up time.

Related issues:
-   https://github.com/mmtk/mmtk-core/issues/669
-   https://github.com/mmtk/mmtk-ruby/issues/13

The problem is, Rust has specialised Vec creation mechanisms for built-in types, such as `usize`.  This means `vec![0usize; 33554432]` will use pre-zeroed buffers to create the Vec, and will not store `0` values to its elements.  This is very fast.  However, Rust cannot do the same optimisation for `vec![SpaceDescriptor(0); 33554432]`. It will have to store a `SpaceDescriptor(0)` to each of its 33554432 elements, and that'll be very slow.

This PR implements our own Vec creation function using `std::alloc::alloc_zeroed`.  This change alone can reduce the execution time of `MMTK::new()` from 140ms to 60ms on my machine.