Other changes related to Immix for Ruby (#754)
These are small changes related to getting Ruby working with Immix.

**Implemented `get_forwarded_object` for ImmixSpace.** `mmtk-core`
didn't use the `get_forwarded_object` method. In theory, we can let Ruby
support copying GC without that method. The `trace_object` method is
sufficient to forward all fields. However, it feels odd to have that
method exposed as a method of `ObjectReference` but not implemented for
Immix. It is a nice debugging tool. It is also useful during weak
reference processing where the VM wants to forward weak fields, but not
expand the transitive closure.

**Removed an assertion that the plan must be non-moving when scanning
non-moving roots.** Now that we already have object pinning for Immix,
it is no longer relevant.