Update dependencies versions (#756)
Some dependencies have new versions. It is a good chance to bump the
dependency versions, too, after we bump the Rust toolchain versions.

One notable dependency is `enum-map`. Its latest version is 2.4.2, but
we locked its version to 2.1.0 because it required a newer Rust
toolchain. Now we can depend on its latest version, instead. We also
changed the `rust-version` property in `Cargo.toml` to `1.61.0` because
`enum-map-derive` depends on that, and `1.61.0` is not new compared to
the `1.66.1` we use. See:
https://github.com/mmtk/mmtk-core/pull/507#discussion_r879081998

This closes https://github.com/mmtk/mmtk-core/issues/693