Introduce immix_non_moving to replace the current immix_no_defrag and immix_no_nursery_copy (#781)
This PR introduces a feature `immix_non_moving`. Based on the proposal
https://github.com/mmtk/mmtk-core/issues/775, defrag or nursery copy
could simply be a constant controlled by the feature `immix_non_moving`.
This also solves the confusion that a binding needs to enable
`immix_no_nursery_copy` to make the Immix policy non-moving even if they
use Immix which does not do nursery copy.